Refactor extended components’s folder structure
- Move all components from src/components/shared/extended to src/components/composite
- Update paths to import statements
