+++ src/screens/Users/screens/Profile/UI.js
@@ -7,7 +7,7 @@ import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
 import { Body, Document, Head, Title } from '../../../../components/common/Document';
 import Layout from '../../../../components/common/Layout';
 import Spinner from '../../../../components/common/Spinner';
-import Error from '../../../../components/shared/extended/Error';
+import Error from '../../../../components/composite/Error';
 
 // Constants
 import PROP_TYPES from '../../../../constants/models/propTypes';
