+++ src/screens/Users/screens/Profile/components/Account.js
@@ -4,7 +4,7 @@ import React from 'react';
 import { Button } from '../../../../../components/common/Buttons';
 import { Card, CardBody, CardHeader, CardSubtitle } from '../../../../../components/common/Card';
 import Text from '../../../../../components/common/Text';
-import Confirm from '../../../../../components/shared/extended/Confirm';
+import Confirm from '../../../../../components/composite/Confirm';
 import stringHelper from '../../../../../helpers/string';
 import timestampHelper from '../../../../../helpers/timestamp';
